Re-worked the set up of the integration tests, so I don't miss anything.
